Moving non Compose classes to commonMain
