+++ writeopia/src/androidMain/kotlin/io/writeopia/sdk/WriteopiaEditor.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import io.writeopia.sdk.draganddrop.target.DraggableScreen
-import io.writeopia.sdk.drawer.DrawInfo
+import io.writeopia.sdk.model.draw.DrawInfo
 import io.writeopia.sdk.drawer.StoryStepDrawer
 import io.writeopia.sdk.model.story.DrawState
