+++ writeopia/src/androidMain/kotlin/io/writeopia/sdk/drawer/commands/CommandsDecoratorDrawer.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import io.writeopia.sdk.drawer.DrawInfo
+import io.writeopia.sdk.model.draw.DrawInfo
 import io.writeopia.sdk.drawer.StoryStepDrawer
 import io.writeopia.sdk.model.action.Action
 import io.writeopia.sdk.models.story.StoryStep
