+++ writeopia/src/androidMain/kotlin/io/writeopia/sdk/drawer/content/SimpleMessageDrawer.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import io.writeopia.sdk.drawer.DrawInfo
+import io.writeopia.sdk.model.draw.DrawInfo
 import io.writeopia.sdk.drawer.StoryStepDrawer
 import io.writeopia.sdk.drawer.modifier.callOnEmptyErase
 import io.writeopia.sdk.model.action.Action
