+++ writeopia/src/commonMain/kotlin/io/writeopia/sdk/model/draw/DrawInfo.kt
@@ -1,4 +1,4 @@
-package io.writeopia.sdk.drawer
+package io.writeopia.sdk.model.draw
 
 /**
  * The class holds the information of the content to be draw by the SDK.
